refactor: implement service layer and component extraction
Backend:
- Create AuthService and UserService for clean architecture
- Extract lifespan logic to app/core/lifespan.py
- Add custom exception handling
- Add missing DELETE endpoints for avatar/cover
- Fix 143 broken tests (335/338 passing - 99% success rate)
- Update test mocks for new service layer

Frontend:
- Extract Dashboard into 7 focused components
- Create reusable hooks (useClock) and utils
- Refactor Navbar with Tailwind CSS
- Extract MobileMenu component

Tests: 335 passed, 3 failed (audit log format only)
