+++ root/app/test-static/avatars/test-avatar-3490b2c229524229ac8f1e1704dd59eb.png
@@ -1,0 +1,1 @@
+avatar
