+++ root/app/test-static/avatars/test-avatar-5942c53d884e4e3d861d90a1a2c9c862.png
@@ -1,0 +1,1 @@
+avatar
