+++ root/app/test-static/covers/test-cover-a88584c0fb1a44538f0cddeda3b0673b.png
@@ -1,0 +1,1 @@
+cover
