+++ root/app/test-static/covers/test-cover-aa913d9617d14fe68fec2ef61943271d.png
@@ -1,0 +1,1 @@
+cover
