+++ root/frontend/src/components/dashboard/ClockWidget.tsx
@@ -1,0 +1,38 @@
+import { Card, Typography } from "@mui/material";
+import { useEffect, useState } from "react";
+import { useTranslation } from "react-i18next";
+
+export function ClockWidget() {
+    const { t } = useTranslation();
+    const [time, setTime] = useState(new Date());
+
+    useEffect(() => {
+        const timer = setInterval(() => setTime(new Date()), 1000);
+        return () => clearInterval(timer);
+    }, []);
+
+    return (
+        <Card
+            sx={{
+                height: "100%",
+                display: "flex",
+                flexDirection: "column",
+                justifyContent: "center",
+                alignItems: "center",
+                p: 2,
+            }}
+        >
+            <Typography variant="h3" component="div" gutterBottom>
+                {time.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" })}
+            </Typography>
+            <Typography variant="subtitle1" color="text.secondary">
+                {time.toLocaleDateString(undefined, {
+                    weekday: "long",
+                    year: "numeric",
+                    month: "long",
+                    day: "numeric",
+                })}
+            </Typography>
+        </Card>
+    );
+}
